+++ jdaviz/core/template_mixin.py
@@ -4418,6 +4418,28 @@ class DatasetSelect(SelectPluginComponent):
         return self.get_object(cls=self.default_data_cls)
 
     def get_selected_spectrum(self, use_display_units=True):
+        """
+        Get the selected spectrum, optionally converted to display units.
+
+        Parameters
+        ----------
+        use_display_units : bool, optional
+            Whether to convert the spectrum to the current display units set in
+            the unit conversion plugin. Default is True.
+
+        Returns
+        -------
+        spectrum : `~specutils.Spectrum1D`
+            The selected spectrum, converted to display units if requested.
+
+        Notes
+        -----
+        If the spectral axis unit of data is pixels, and the
+        display unit is not pixels (or vice versa), no conversion is done to allow
+        for mixed pixel/world unit viewing (this logic is handled by
+        spectral_axis_conversion, which is called from this method when converting
+        the spectral axis).
+        """
         # retrieves the 1d spectrum
         if isinstance(self.selected_obj, NDData):
             shape = self.selected_obj.data.shape